review marc - renaming TerminLoeschen to DeleteTerminConfirmation
