+++ src/containers/DeleteTerminConfirmation.css
@@ -1,3 +1,3 @@
-.LoeschenDialog {
+.DeleteTerminConfirmation {
     background-color: #edcfb7;
 }
